make default constractor ex0*

+++ ex02/ClapTrap.cpp
@@ -69,7 +69,7 @@ void ClapTrap::beRepaired(unsigned int amount){
 		return ;
 	}
 	std::cout << _name << " repaired " << amount << " of hp" << std::endl;
-	if (INT_MAX - _hp <= amount)
+	if (static_cast<unsigned int>(INT_MAX - _hp) <= amount)
 		_hp = INT_MAX;
 	else
 		_hp += amount;
